Support SCOPE_VER and Sideloading
This adds support for setting the SCOPE_VER envrionment variable when
running start.sh to specify the image tag for cribl/scope to use like
so.

    $ SCOPE_VER=0.7.0-rc5 ./start.sh cribl

This also adds a step in start.sh to sideload the cribl/scope image into
the cluster if it's found among the local Docker images. For debugging.